requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
appnope==0.1.2
APScheduler==3.8.1
astroid==2.4.2
async-timeout==4.0.0
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cycler==0.10.0
dateparser==1.1.0
decorator==4.4.2
et-xmlfile==1.0.1
Flask==2.0.2
frozenlist==1.2.0
idna==2.10
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.2
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
multidict==5.2.0
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pygame==2.0.1
Pygments==2.8.0
pylint==2.6.0
pyparsing==2.4.7
python-binance==1.0.15
python-dateutil==2.8.1
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.0.3
regex==2021.11.1
requests==2.25.1
schedule==1.1.0
selenium==3.141.0
six==1.15.0
songline==0.0.1
soupsieve==2.2
TA-Lib==0.4.21
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
unclechat==0.0.1
urllib3==1.26.3
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.2
wrapt==1.12.1
xlrd==2.0.1
yarl==1.7.2
